+++ mame/src/lib/formats/esq16_dsk.h
@@ -1,3 +1,5 @@
+// license:BSD-3-Clause
+// copyright-holders:R. Belmont, Olivier Galibert
 /*********************************************************************
 
     formats/esq16_dsk.h
@@ -30,7 +32,7 @@ public:
 	static const desc_e esq_10_desc[];
 
 private:
-	void find_size(io_generic *io, int &track_count, int &head_count, int &sector_count);
+	void find_size(io_generic *io, UINT8 &track_count, UINT8 &head_count, UINT8 &sector_count);
 };
 
 extern const floppy_format_type FLOPPY_ESQIMG_FORMAT;
