+++ mame/src/lib/formats/st_dsk.h
@@ -1,3 +1,5 @@
+// license:BSD-3-Clause
+// copyright-holders:Olivier Galibert
 /*********************************************************************
 
     formats/st_dsk.h
@@ -26,7 +28,7 @@ public:
 	virtual bool supports_save() const;
 
 private:
-	void find_size(io_generic *io, int &track_count, int &head_count, int &sector_count);
+	void find_size(io_generic *io, UINT8 &track_count, UINT8 &head_count, UINT8 &sector_count);
 };
 
 class msa_format : public floppy_image_format_t
